ipcache/md: Move host label logic to injectLabels
Move the logic to change the host label into injectLabels() so that the
caller doesn't need to worry about this detail. This allows injectLabels
to fully handle ensuring that the specified labels are present in the
returned identity, which will simplify further upcoming changes to
generalize this code. In particular it will make it easier to handle
identity (re-)allocations for other label-based features.

Signed-off-by: Joe Stringer <joe@cilium.io>